+++ test.cub
@@ -9,9 +9,9 @@ C  124,185,232
 11100000000111
 10000000000001
 11011000000001
-11000000111101
+1100000E111101
 10000000000011
-1000000N000001
+10000000000001
 10000000000001
 11000000000001
 10000000000001
